pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "megaton"
version = "1.0.1"
description = "Python toolkit for running GA4 and Search Console reports in Colab/Notebooks and publishing results to Google Sheets, CSV, or BigQuery, with built-in retries and easy append/upsert."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Makoto Shimizu", email = "aa.analyst.ga@gmail.com" },
]
keywords = [
    "google-analytics", "ga4", "search-console", "google-sheets",
    "bigquery", "jupyter", "notebook", "colab",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Internet",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "pandas>=1.3",
    "google-api-python-client>=2.89",
    "google-auth>=2.40.0",
    "google-api-core>=2.25.0",
    "google-auth-oauthlib>=1.0",
    "google-analytics-data>=0.18",
    "google-analytics-admin>=0.25.0",
    "google-cloud-bigquery>=3.11",
    "gspread>=5.7",
    "oauthlib>=3.2",
    "ipywidgets>=8.0",
    "gspread-dataframe>=3.0",
    "pytz>=2023.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[project.urls]
Homepage = "https://github.com/mak00s/megaton"
Repository = "https://github.com/mak00s/megaton"
Documentation = "https://github.com/mak00s/megaton/tree/main/docs"
Issues = "https://github.com/mak00s/megaton/issues"
Changelog = "https://github.com/mak00s/megaton/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
include = ["megaton*"]

[tool.setuptools.package-data]
megaton = ["py.typed"]

[tool.coverage.run]
source = ["megaton"]
omit = [
    "megaton/ga3.py",
]
